More work on collision between ball and paddle
The ball now takes a percentage of the paddles current y velocity as
it's own, adding more to the gameplay. This is not finished however, and
paddles will get stuck to the bottom or top of the screen if they touch.
Started a little but of work on buttons which will be used later in a
menu, and also added controls to the README.